fix signed buffer size no longer causes a segmentation fault
Call of pwgen --no-vowels $(( 2**31 )) 1 no longer causes a segmentation fault

+++ pwgen.h
@@ -7,6 +7,7 @@
  * License.
  */
 
+#include <stddef.h>
 struct pw_element {
 	const char	*str;
 	int		flags;
@@ -38,10 +39,10 @@ extern const char *pw_ambiguous;
 /* Function prototypes */
 
 /* pw_phonemes.c */
-extern void pw_phonemes(char *buf, int size, int pw_flags, char *remove);
+extern void pw_phonemes(char *buf, size_t size, int pw_flags, char *remove);
 
 /* pw_rand.c */
-extern void pw_rand(char *buf, int size, int pw_flags, char *remove);
+extern void pw_rand(char *buf, size_t size, int pw_flags, char *remove);
 
 /* randnum.c */
 extern int pw_random_number(int max_num);
